runtime: Deprecate transaction::MethodDispatcher
By deprecate I mean, move into the simple-keyvalue runtime's main.rs,
preserving existing functionality in the worst possible way.
