+++ .changelog/4198.breaking.md
@@ -1,0 +1,1 @@
+runtime: Deprecate transaction::MethodDispatcher
