+++ tests/runtimes/simple-keyvalue/Cargo.toml
@@ -20,6 +20,7 @@ stack-size = 2097152
 threads = 2
 
 [dependencies]
+cbor = { version = "0.1.0", package = "oasis-cbor" }
 oasis-core-runtime = { path = "../../../runtime" }
 oasis-core-client = { path = "../../../client" }
 oasis-core-keymanager-client = { path = "../../../keymanager-client" }
@@ -29,6 +30,7 @@ simple-keyvalue-api = { path = "./api" }
 
 # Third party.
 anyhow = "1.0"
+thiserror = "1.0"
 io-context = "0.2.0"
 byteorder = "1.4.3"
