+++ tests/runtimes/simple-keyvalue/src/main.rs
@@ -1,11 +1,19 @@
 // Allow until oasis-core#3572.
 #![allow(deprecated)]
 
-use std::{collections::BTreeMap, io::Cursor, sync::Arc};
+use std::{
+    collections::{BTreeMap, HashMap},
+    io::Cursor,
+    sync::{
+        atomic::{AtomicBool, Ordering},
+        Arc,
+    },
+};
 
-use anyhow::{anyhow, Result};
+use anyhow::{anyhow, Context as AnyContext, Result};
 use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
 use io_context::Context as IoContext;
+use thiserror::Error;
 
 use oasis_core_keymanager_client::{KeyManagerClient, KeyPairId};
 use oasis_core_runtime::{
@@ -26,18 +34,410 @@ use oasis_core_runtime::{
         state::staking::ImmutableState as StakingImmutableState,
     },
     rak::RAK,
-    register_runtime_txn_methods, runtime_context,
+    runtime_context,
     storage::{StorageContext, MKVS},
     transaction::{
-        dispatcher::{BatchHandler, CheckOnlySuccess},
+        dispatcher::{Dispatcher, ExecuteBatchResult, ExecuteTxResult},
+        tags::Tags,
+        types::{TxnBatch, TxnCall, TxnCheckResult, TxnOutput},
         Context as TxnContext,
     },
-    version_from_cargo, Protocol, RpcDemux, RpcDispatcher, TxnDispatcher, TxnMethDispatcher,
+    types::{CheckTxResult, Error as RuntimeError},
+    version_from_cargo, Protocol, RpcDemux, RpcDispatcher, TxnDispatcher,
 };
 use simple_keymanager::trusted_policy_signers;
 use simple_keyvalue_api::{
     with_api, AddEscrow, Key, KeyValue, ReclaimEscrow, Transfer, UpdateRuntime, Withdraw,
 };
+
+// This is the old runtime method dispatcher that used to be part
+// of the oasis-core/runtime API.  It has been moved here since
+// new runtimes should use the SDK instead.
+
+/// Registers defined transaction methods into the transaction dispatcher.
+///
+/// # Examples
+///
+/// This macro should be invoked using a concrete API generated by `runtime_api`
+/// as follows:
+/// ```rust,ignore
+/// with_api! {
+///     register_runtime_txn_methods!(txn_dispatcher, api);
+/// }
+/// ```
+macro_rules! register_runtime_txn_methods {
+    (
+        $txn_dispatcher:ident,
+        $(
+            pub fn $method_name:ident ( $arguments_type:ty ) -> $output_type:ty ;
+        )*
+    ) => {
+        $(
+            $txn_dispatcher.add_method(
+               Method::new(
+                    MethodDescriptor {
+                        name: stringify!($method_name).to_owned(),
+                    },
+                    |args: &$arguments_type,
+                     ctx: &mut oasis_core_runtime::transaction::context::Context|
+                        -> ::anyhow::Result<$output_type> {
+                        $method_name(args, ctx)
+                    },
+                )
+            );
+        )*
+    }
+}
+
+/// Dispatch error.
+#[derive(Error, Debug)]
+enum DispatchError {
+    #[error("method not found: {method:?}")]
+    MethodNotFound { method: String },
+}
+
+/// Error indicating that performing a transaction check was successful.
+#[derive(Error, Debug, Default)]
+#[error("transaction check successful")]
+pub struct CheckOnlySuccess(pub TxnCheckResult);
+
+/// Custom batch handler.
+///
+/// A custom batch handler can be configured on the `Dispatcher` and will have
+/// its `start_batch` and `end_batch` methods called at the appropriate times.
+pub trait BatchHandler {
+    /// Called before the first call in a batch is dispatched.
+    ///
+    /// The context may be mutated and will be available as read-only to all
+    /// runtime calls.
+    fn start_batch(&self, ctx: &mut TxnContext);
+
+    /// Called after all calls have been dispatched.
+    fn end_batch(&self, ctx: &mut TxnContext);
+}
+
+/// Custom context initializer.
+pub trait ContextInitializer {
+    /// Called to initialize the context.
+    fn init(&self, ctx: &mut TxnContext);
+}
+
+impl<F> ContextInitializer for F
+where
+    F: Fn(&mut TxnContext),
+{
+    fn init(&self, ctx: &mut TxnContext) {
+        (*self)(ctx)
+    }
+}
+
+/// Custom finalizer.
+pub trait Finalizer {
+    /// Called to finalize transaction.
+    ///
+    /// This method is called after storage has been finalized so the
+    /// storage context is not available and using it will panic.
+    fn finalize(&self, new_storage_root: Hash);
+}
+
+impl<F> Finalizer for F
+where
+    F: Fn(Hash),
+{
+    fn finalize(&self, new_storage_root: Hash) {
+        (*self)(new_storage_root)
+    }
+}
+
+/// Descriptor of a runtime API method.
+#[derive(Clone, Debug)]
+pub struct MethodDescriptor {
+    /// Method name.
+    pub name: String,
+}
+
+/// Handler for a runtime method.
+pub trait MethodHandler<Call, Output> {
+    /// Invoke the method implementation and return a response.
+    fn handle(&self, call: &Call, ctx: &mut TxnContext) -> Result<Output>;
+}
+
+impl<Call, Output, F> MethodHandler<Call, Output> for F
+where
+    Call: 'static,
+    Output: 'static,
+    F: Fn(&Call, &mut TxnContext) -> Result<Output> + 'static,
+{
+    fn handle(&self, call: &Call, ctx: &mut TxnContext) -> Result<Output> {
+        (*self)(&call, ctx)
+    }
+}
+
+/// Dispatcher for a runtime method.
+pub trait MethodHandlerDispatch {
+    /// Get method descriptor.
+    fn get_descriptor(&self) -> &MethodDescriptor;
+
+    /// Dispatches the given raw call.
+    fn dispatch(&self, call: TxnCall, ctx: &mut TxnContext) -> Result<cbor::Value>;
+}
+
+struct MethodHandlerDispatchImpl<Call, Output> {
+    /// Method descriptor.
+    descriptor: MethodDescriptor,
+    /// Method handler.
+    handler: Box<dyn MethodHandler<Call, Output>>,
+}
+
+impl<Call, Output> MethodHandlerDispatch for MethodHandlerDispatchImpl<Call, Output>
+where
+    Call: cbor::Decode + 'static,
+    Output: cbor::Encode + 'static,
+{
+    fn get_descriptor(&self) -> &MethodDescriptor {
+        &self.descriptor
+    }
+
+    fn dispatch(&self, call: TxnCall, ctx: &mut TxnContext) -> Result<cbor::Value> {
+        let call = cbor::from_value(call.args).context("unable to parse call arguments")?;
+        let response = self.handler.handle(&call, ctx)?;
+
+        Ok(cbor::to_value(response))
+    }
+}
+
+/// Runtime method dispatcher implementation.
+pub struct Method {
+    /// Method dispatcher.
+    dispatcher: Box<dyn MethodHandlerDispatch>,
+}
+
+impl Method {
+    /// Create a new enclave method descriptor.
+    pub fn new<Call, Output, Handler>(method: MethodDescriptor, handler: Handler) -> Self
+    where
+        Call: cbor::Decode + 'static,
+        Output: cbor::Encode + 'static,
+        Handler: MethodHandler<Call, Output> + 'static,
+    {
+        Method {
+            dispatcher: Box::new(MethodHandlerDispatchImpl {
+                descriptor: method,
+                handler: Box::new(handler),
+            }),
+        }
+    }
+
+    /// Return method name.
+    pub fn get_name(&self) -> &String {
+        &self.dispatcher.get_descriptor().name
+    }
+
+    /// Dispatch method call.
+    pub fn dispatch(&self, call: TxnCall, ctx: &mut TxnContext) -> Result<cbor::Value> {
+        self.dispatcher.dispatch(call, ctx)
+    }
+}
+
+/// Runtime method dispatcher.
+///
+/// The dispatcher is a concrete implementation of the Dispatcher trait.
+/// It holds all registered runtime methods and provides an entry point
+/// for their invocation.
+pub struct MethodDispatcher {
+    /// Registered runtime methods.
+    methods: HashMap<String, Method>,
+    /// Registered batch handler.
+    batch_handler: Option<Box<dyn BatchHandler>>,
+    /// Registered context initializer.
+    ctx_initializer: Option<Box<dyn ContextInitializer>>,
+    /// Registered finalizer.
+    finalizer: Option<Box<dyn Finalizer>>,
+    /// Abort batch flag.
+    abort_batch: Option<Arc<AtomicBool>>,
+}
+
+impl MethodDispatcher {
+    /// Create a new runtime method dispatcher instance.
+    pub fn new() -> MethodDispatcher {
+        MethodDispatcher {
+            methods: HashMap::new(),
+            batch_handler: None,
+            ctx_initializer: None,
+            finalizer: None,
+            abort_batch: None,
+        }
+    }
+
+    /// Register a new method in the dispatcher.
+    pub fn add_method(&mut self, method: Method) {
+        self.methods.insert(method.get_name().clone(), method);
+    }
+
+    /// Configure batch handler.
+    pub fn set_batch_handler<H>(&mut self, handler: H)
+    where
+        H: BatchHandler + 'static,
+    {
+        self.batch_handler = Some(Box::new(handler));
+    }
+
+    /// Configure context initializer.
+    pub fn set_context_initializer<I>(&mut self, initializer: I)
+    where
+        I: ContextInitializer + 'static,
+    {
+        self.ctx_initializer = Some(Box::new(initializer));
+    }
+
+    /// Configure finalizer.
+    #[allow(dead_code)]
+    pub fn set_finalizer<F>(&mut self, finalizer: F)
+    where
+        F: Finalizer + 'static,
+    {
+        self.finalizer = Some(Box::new(finalizer));
+    }
+
+    /// Dispatches a raw runtime check request.
+    fn dispatch_check(&self, call: &Vec<u8>, ctx: &mut TxnContext) -> CheckTxResult {
+        match self.dispatch_fallible(call, ctx) {
+            Ok(_response) => CheckTxResult {
+                error: Default::default(),
+                // Deprecated method dispatcher doesn't support check tx metadata.
+                meta: None,
+            },
+            Err(error) => match error.downcast::<CheckOnlySuccess>() {
+                Ok(_check_result) => CheckTxResult {
+                    error: Default::default(),
+                    // Deprecated method dispatcher doesn't support check tx metadata.
+                    meta: None,
+                },
+                Err(error) => CheckTxResult {
+                    error: RuntimeError {
+                        module: "".to_string(),
+                        code: 1,
+                        message: format!("{:#}", error),
+                    },
+                    meta: None,
+                },
+            },
+        }
+    }
+
+    /// Dispatches a raw runtime invocation request.
+    fn dispatch_execute(&self, call: &Vec<u8>, ctx: &mut TxnContext) -> ExecuteTxResult {
+        let rsp = match self.dispatch_fallible(call, ctx) {
+            Ok(response) => TxnOutput::Success(response),
+            Err(error) => TxnOutput::Error(format!("{:#}", error)),
+        };
+
+        ExecuteTxResult {
+            output: cbor::to_vec(rsp),
+            tags: ctx.take_tags(),
+        }
+    }
+
+    fn dispatch_fallible(&self, call: &Vec<u8>, ctx: &mut TxnContext) -> Result<cbor::Value> {
+        let call: TxnCall = cbor::from_slice(call).context("unable to parse call")?;
+
+        match self.methods.get(&call.method) {
+            Some(dispatcher) => dispatcher.dispatch(call, ctx),
+            None => Err(DispatchError::MethodNotFound {
+                method: call.method,
+            }
+            .into()),
+        }
+    }
+}
+
+impl Dispatcher for MethodDispatcher {
+    fn check_batch(
+        &self,
+        mut ctx: TxnContext,
+        batch: &TxnBatch,
+    ) -> Result<Vec<CheckTxResult>, RuntimeError> {
+        if let Some(ref ctx_init) = self.ctx_initializer {
+            ctx_init.init(&mut ctx);
+        }
+
+        // Invoke start batch handler.
+        if let Some(ref handler) = self.batch_handler {
+            handler.start_batch(&mut ctx);
+        }
+
+        // Process batch.
+        let mut results = Vec::new();
+        for call in batch.iter() {
+            if self
+                .abort_batch
+                .as_ref()
+                .map(|b| b.load(Ordering::SeqCst))
+                .unwrap_or(false)
+            {
+                return Err(RuntimeError::new("rhp/dispatcher", 1, "batch aborted"));
+            }
+            results.push(self.dispatch_check(call, &mut ctx));
+            let _ = ctx.take_tags();
+        }
+
+        Ok(results)
+    }
+
+    fn execute_batch(
+        &self,
+        mut ctx: TxnContext,
+        batch: &TxnBatch,
+    ) -> Result<ExecuteBatchResult, RuntimeError> {
+        if let Some(ref ctx_init) = self.ctx_initializer {
+            ctx_init.init(&mut ctx);
+        }
+
+        // Invoke start batch handler.
+        if let Some(ref handler) = self.batch_handler {
+            handler.start_batch(&mut ctx);
+        }
+
+        // Process batch.
+        let mut results = Vec::new();
+        for call in batch.iter() {
+            if self
+                .abort_batch
+                .as_ref()
+                .map(|b| b.load(Ordering::SeqCst))
+                .unwrap_or(false)
+            {
+                return Err(RuntimeError::new("rhp/dispatcher", 1, "batch aborted"));
+            }
+            results.push(self.dispatch_execute(call, &mut ctx));
+        }
+
+        // Invoke end batch handler.
+        if let Some(ref handler) = self.batch_handler {
+            handler.end_batch(&mut ctx);
+        }
+
+        Ok(ExecuteBatchResult {
+            results,
+            messages: ctx.close(),
+            // No support for block tags in the deprecated dispatcher.
+            block_tags: Tags::new(),
+            // No support for custom batch weight limits.
+            batch_weight_limits: None,
+        })
+    }
+
+    fn finalize(&self, new_storage_root: Hash) {
+        if let Some(ref finalizer) = self.finalizer {
+            finalizer.finalize(new_storage_root);
+        }
+    }
+
+    /// Configure abort batch flag.
+    fn set_abort_batch_flag(&mut self, abort_batch: Arc<AtomicBool>) {
+        self.abort_batch = Some(abort_batch);
+    }
+}
 
 /// Key format used for transaction artifacts.
 #[derive(Debug)]
@@ -586,7 +986,7 @@ pub fn main() {
                 _rpc_demux: &mut RpcDemux,
                 rpc: &mut RpcDispatcher|
      -> Option<Box<dyn TxnDispatcher>> {
-        let mut txn = TxnMethDispatcher::new();
+        let mut txn = MethodDispatcher::new();
         with_api! { register_runtime_txn_methods!(txn, api); }
 
         // Create the key manager client.
